+++ shared_network/src/Messages/TextMessage.cpp
@@ -11,7 +11,7 @@ TextMessage::TextMessage(char* buffer) : Message(buffer)
 TextMessage::TextMessage(const std::string& text, TextType textType, unsigned int senderID)
 {
 	header.type = MessageType::TEXT;
-	header.size = sizeof(TextType) + sizeof(uint8_t) + (uint8_t)text.size();
+	header.size = sizeof(TextType) + sizeof(uint8_t) + text.size();
 	header.id = senderID;
 
 	//copy data to message data
